# docs/swagger.yaml
definitions:
  github_com_maidulcu_masaar-crm_internal_domain.Contact:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      language:
        description: '"ar" | "en"'
        type: string
      lead_score:
        type: integer
      phone_wa:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_maidulcu_masaar-crm_internal_domain.Lead:
    properties:
      contact:
        allOf:
        - $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.Contact'
        description: Joined
      contact_id:
        type: string
      created_at:
        type: string
      currency:
        description: 'default: AED'
        type: string
      deal_value:
        type: number
      id:
        type: string
      notes:
        type: string
      source:
        $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.LeadSource'
      stage:
        $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.LeadStage'
      updated_at:
        type: string
    type: object
  github_com_maidulcu_masaar-crm_internal_domain.LeadSource:
    enum:
    - whatsapp
    - web
    - referral
    - event
    type: string
    x-enum-varnames:
    - SourceWhatsApp
    - SourceWeb
    - SourceReferral
    - SourceEvent
  github_com_maidulcu_masaar-crm_internal_domain.LeadStage:
    enum:
    - new
    - contacted
    - qualified
    - proposal
    - won
    - lost
    type: string
    x-enum-varnames:
    - StageNew
    - StageContacted
    - StageQualified
    - StageProposal
    - StageWon
    - StageLost
...
info:
  contact: {}
paths:
...
  /leads:
    get:
      description: Returns all leads grouped by stage for the Kanban pipeline view.
      produces:
      - application/json
      responses:
        "200":
          description: Map of stage → []Lead
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get Kanban board
      tags:
      - Leads
    post:
      consumes:
      - application/json
      description: Creates a new lead and broadcasts a lead.created WebSocket event.
      parameters:
      - description: Lead payload (contact_id required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.Lead'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.Lead'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create lead
      tags:
      - Leads
  /leads/{id}:
    get:
      description: Returns a single lead with its contact joined.
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.Lead'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get lead
      tags:
      - Leads
  /leads/{id}/stage:
    patch:
      consumes:
      - application/json
      description: Updates the pipeline stage of a lead (drag-drop). Broadcasts lead.stage_changed
        event.
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      - description: 'New stage: new|contacted|qualified|proposal|won|lost'
        in: body
        name: body
        required: true
        schema:
          properties:
            stage:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              lead_id:
                type: string
              stage:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move lead to stage
      tags:
      - Leads
...
swagger: "2.0"